+++ src/benchmark/relative_speed.rs
@@ -136,6 +136,7 @@ fn create_result(name: &str, mean: Scalar) -> BenchmarkResult {
         memory_usage_byte: None,
         exit_codes: Vec::new(),
         parameters: BTreeMap::new(),
+        ..Default::default()
     }
 }