+++ src/benchmark/timing_result.rs
@@ -14,4 +14,14 @@ pub struct TimingResult {
 
     /// Maximum amount of memory used, in bytes
     pub memory_usage_byte: u64,
+    /// number of voluntary context swithcing
+    pub voluntary_cs: u64,
+    /// number of involuntary context swithcing
+    pub involuntary_cs: u64,
+
+    ///  number of io read operations
+    pub io_read_ops: u64,
+
+    /// number of io write operations
+    pub io_write_ops: u64,
 }