+++ src/timer/mod.rs
@@ -12,6 +12,8 @@ use nix::fcntl::{splice, SpliceFFlags};
 use std::fs::File;
 #[cfg(target_os = "linux")]
 use std::os::fd::AsFd;
+#[cfg(unix)]
+use std::os::unix::process::ExitStatusExt;
 
 #[cfg(target_os = "windows")]
 use windows_sys::Win32::System::Threading::CREATE_SUSPENDED;
@@ -44,6 +46,10 @@ pub struct TimerResult {
     pub time_user: Second,
     pub time_system: Second,
     pub memory_usage_byte: u64,
+    pub voluntary_cs: u64,
+    pub involuntary_cs: u64,
+    pub io_read_ops: u64,
+    pub io_write_ops: u64,
     /// The exit status of the process
     pub status: ExitStatus,
 }
@@ -87,35 +93,91 @@ pub fn execute_and_measure(mut command: Command) -> Result<TimerResult> {
     #[cfg(windows)]
     {
         use std::os::windows::process::CommandExt;
-
-        // Create the process in a suspended state so that we don't miss any cpu time between process creation and `CPUTimer` start.
         command.creation_flags(CREATE_SUSPENDED);
     }
 
     let wallclock_timer = WallClockTimer::start();
-    let mut child = command.spawn()?;
+    let mut child = command.spawn()?; // <-- Start process
 
     #[cfg(windows)]
-    let cpu_timer = {
-        // SAFETY: We created a suspended process
-        unsafe { self::windows_timer::CPUTimer::start_suspended_process(&child) }
-    };
+    let cpu_timer = { unsafe { self::windows_timer::CPUTimer::start_suspended_process(&child) } };
 
     if let Some(output) = child.stdout.take() {
-        // Handle CommandOutputPolicy::Pipe
         discard(output);
     }
+    #[cfg(unix)]
+    let (status, io_read, io_writes, voluntary_cs, involuntary_cs) = {
+        use libc::{c_int, rusage};
 
-    let status = child.wait()?;
+        let pid = child.id() as libc::pid_t;
+
+        // Wait for child and get rusage
+        let mut status: c_int = 0;
+        let mut usage: rusage = unsafe { std::mem::zeroed() };
+
+        unsafe {
+            use libc::wait4;
+
+            wait4(pid, &mut status, 0, &mut usage);
+        }
+        (
+            ExitStatus::from_raw(status),
+            usage.ru_inblock,
+            usage.ru_oublock,
+            usage.ru_nvcsw,
+            usage.ru_nivcsw,
+        )
+    };
+
+    // Note that you can not use both child.wait()? and wait4
+    // Because both pull the process from the process table
+    #[cfg(windows)]
+    let status = child.wait()?; // <-- Wait for completion
 
     let time_real = wallclock_timer.stop();
     let (time_user, time_system, memory_usage_byte) = cpu_timer.stop();
+
+    //  Collect extra stats on Windows
+    #[cfg(windows)]
+    let (voluntary_cs, involuntary_cs, io_read, io_writes) = {
+        use std::mem::MaybeUninit;
+        use std::os::windows::io::AsRawHandle;
+        use windows_sys::Win32::{
+            Foundation::HANDLE, System::Threading::GetProcessIoCounters,
+            System::Threading::IO_COUNTERS,
+        };
+
+        let handle: HANDLE = child.as_raw_handle() as HANDLE;
+        let mut counters = MaybeUninit::<IO_COUNTERS>::uninit();
+
+        let success = unsafe { GetProcessIoCounters(handle, counters.as_mut_ptr()) };
+        if success == 0 {
+            (0, 0, 0, 0)
+        } else {
+            let counters = unsafe { counters.assume_init() };
+            (
+                0, // Can not find any API  available for Context  switches  in windows
+                0,
+                counters.ReadOperationCount as u64,
+                counters.WriteOperationCount as u64,
+            )
+        }
+    };
+
+    let voluntary_cs = voluntary_cs as u64;
+    let involuntary_cs = involuntary_cs as u64;
+    let io_read_ops = io_read as u64;
+    let io_write_ops = io_writes as u64;
 
     Ok(TimerResult {
         time_real,
         time_user,
         time_system,
         memory_usage_byte,
+        voluntary_cs,
+        involuntary_cs,
+        io_read_ops,
+        io_write_ops,
         status,
     })
 }